[TBDGen] NFC: move TBDGenVisitor to its own header.
